seperate activity for Database Students
