+++ IntegratedProject/app/src/main/java/com/example/integratedproject/MainActivity.kt
@@ -11,38 +11,9 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        title = "iWasThere"
-        val context = this
-        val db = DataBaseHandler(context)
-        AddUserButton.setOnClickListener {
-            if(editTextUserName.text.toString().isNotEmpty() &&
-                    editTextTextStudentenNummer.text.toString().isNotEmpty()){
-                val name = editTextUserName.text.toString()
-                val studentenNummer = editTextTextStudentenNummer.text.toString()
 
-                db.insertData(name, studentenNummer)
-                clearField()
-            }
-            else{
-                Toast.makeText(context, "Please fill in all fields", Toast.LENGTH_SHORT).show()
-            }
-
-        }
-
-        ButtonViewData.setOnClickListener {
-            val data = db.readData()
-            textViewResult.text = ""
-            for(i in 0 until data.size){
-                textViewResult.append(
-                    data[i].id.toString() + " " + data[i].name + " " + data[i].studentennummer.toString() + "\n"
-                )
-            }
-        }
 
     }
 
-    private fun clearField(){
-        editTextUserName.text.clear()
-        editTextTextStudentenNummer.text.clear()
-    }
+
 }
