+++ IntegratedProject/app/src/main/res/layout-land/activity_main.xml
@@ -6,87 +6,5 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <Button
-        android:id="@+id/ButtonViewData"
-        android:layout_width="155dp"
-        android:layout_height="70dp"
-        android:layout_marginEnd="100dp"
-        android:layout_marginRight="100dp"
-        android:layout_marginBottom="192dp"
-        android:text="View Data"
-        android:textSize="18sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <TextView
-        android:id="@+id/UserNametextView"
-        android:layout_width="135dp"
-        android:layout_height="53dp"
-        android:layout_marginEnd="400dp"
-        android:layout_marginRight="400dp"
-        android:layout_marginBottom="444dp"
-        android:text="naam"
-        android:textSize="30sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <EditText
-        android:id="@+id/editTextUserName"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="376dp"
-        android:layout_marginRight="376dp"
-        android:layout_marginBottom="388dp"
-        android:ems="10"
-        android:inputType="textPersonName"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <TextView
-        android:id="@+id/studentenNummerpassword"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="356dp"
-        android:layout_marginRight="356dp"
-        android:layout_marginBottom="336dp"
-        android:text="studentennummer"
-        android:textSize="30sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <EditText
-        android:id="@+id/editTextTextStudentenNummer"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="376dp"
-        android:layout_marginRight="376dp"
-        android:layout_marginBottom="264dp"
-        android:ems="10"
-        android:inputType="textPassword"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <Button
-        android:id="@+id/AddUserButton"
-        android:layout_width="155dp"
-        android:layout_height="70dp"
-        android:layout_marginEnd="744dp"
-        android:layout_marginRight="744dp"
-        android:layout_marginBottom="192dp"
-        android:text="Add user"
-        android:textSize="18sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <TextView
-        android:id="@+id/textViewResult"
-        android:layout_width="343dp"
-        android:layout_height="127dp"
-        android:layout_marginEnd="294dp"
-        android:layout_marginRight="294dp"
-        android:layout_marginBottom="33dp"
-        android:text="TextView"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
